feat: fix postgresql not found on special backups
pgBackRest does not guarantee that postgresql.conf would exist on a particular backup nor it will create one based on the current server version. Thus, when restoring from a backup that has been created without a config file, we have to backup the one that initdb has generated and restore it (but without overwriting existing files) to the pgdata directory.